requirements.txt
selenium
PyPDF2
pyTelegramBotAPI
telebot
python-telegram-bot